ref(rcs): move commands to spot-remote rcs
